requirements.txt
absl-py==1.2.0
argon2-cffi==21.1.0
attrs==21.2.0
bcrypt==3.2.0
bleach==4.1.0
cachetools==5.2.0
cffi==1.14.6
charset-normalizer==2.1.1
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.37.3
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
Jinja2==3.0.2
joblib==1.2.0
jsonschema==4.1.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
networkx==2.6.3
notebook==6.4.4
numpy==1.21.6
oauthlib==3.2.1
packaging==21.0
pandas==1.3.5
pandocfilters==1.5.0
Pillow==7.1.2
prometheus-client==0.11.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch @ https://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp37-cp37m-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu113/torchaudio-0.12.1%2Bcu113-cp37-cp37m-linux_x86_64.whl
torchsummary==1.5.1
torchtext==0.13.1
torchvision @ https://download.pytorch.org/whl/cu113/torchvision-0.13.1%2Bcu113-cp37-cp37m-linux_x86_64.whl
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.1
zipp==3.8.1
wandb
omegaconf
pytorch-lightning
